Dodanie stylizacji dla todoApp + przeniesienie todoApp do oddzielnego komponentu :)
