+++ package.json
@@ -9,6 +9,7 @@
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
     "react-scripts": "4.0.3",
+    "sweetalert2": "^10.15.7",
     "uuid": "^8.3.2",
     "web-vitals": "^1.1.1"
   },
