+++ src/App.js
@@ -1,54 +1,11 @@
-import { useState, useRef, useEffect } from 'react';
-import TodoList from './components/TodoList';
-import { v4 as uuidv4 } from 'uuid';
-
-const LOCAL_STORAGE_KEY = 'todoApp.todos';
+import TodoApp from './components/TodoApp'
 
 function App() {
-  const [todos, setTodos] = useState([])
-  const todoNameRef = useRef()
 
-  useEffect(() => {
-    const storedTodos = JSON.parse(localStorage.getItem(LOCAL_STORAGE_KEY))
-    if (storedTodos) setTodos(storedTodos)
-  }, [])
-
-  useEffect(() => {
-    localStorage.setItem(LOCAL_STORAGE_KEY, JSON.stringify(todos))
-  }, [todos])
-
-  function toggleTodo(id) {
-    const newTodos = [...todos]
-    const todo = newTodos.find(todo => todo.id === id)
-    todo.complete = !todo.complete
-    setTodos(newTodos)
-  }
-
-  function handleAddTodo(e) {
-    const name = todoNameRef.current.value
-    if (name === '') return
-
-    setTodos(prevTodos => {
-      return [...prevTodos, { id: uuidv4(), name: name, complete: false }]
-    })
-
-    todoNameRef.current.value = null
-  }
-
-  function handleClearTodos() {
-    const newTodos = todos.filter(todo => !todo.complete)
-    setTodos(newTodos)
-  }
 
   return (
     <>
-      <TodoList todos={todos} toggleTodo={toggleTodo} />
-      <form>
-        <input ref={todoNameRef} type="text" required />
-        <button onClick={handleAddTodo}>Add todo</button>
-        <button onClick={handleClearTodos}>Clear Complete</button>
-      </form>
-      <div>{todos.filter(todo => !todo.complete).length} left to do</div>
+      <TodoApp />
     </>
   );
 }
